requirements.txt
kafka-python==1.4.1
pykafka==2.7.0
